# doc/api.yaml
openapi: 3.0.0
info:
  title : WASAphoto 
  description: |
    WASAphoto: uploading and sharing photos with your friends. API made by 1982464 .
  version : 0.0.1

########## CONTROLLARE TUTTI I TAGS PERCHÈ NON HANNO SENSO 
paths:
  /session: 

    post:
      operationId: doLogin 
      tags: ["login"]
      summary: Logs in the user 
      description: |-
        If the user does not exist, it will be created, and returned 
        If the user exists, is returned.

      requestBody:
        description: Nickname
        required: true 
        content:
          application/json: 
            schema: {$ref: "#/components/schemas/Complete_user"}
       
      responses:

        '201':
          description: User CREATED; sing-up action and returned the content 
          content:
            application/json: 
              schema: {$ref: "#/components/schemas/Complete_user"}        
        "200":
          description: User log-in action, returned the content.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Complete_user' }
        
        "400":
          description: Invalid JSON object or invalid username.

        "500":
          # Internal Server Error
          $ref: '#/components/responses/Server_error'
  /user:
    put: 
      operationId: setMyUserName
      tags: [user]
      summary: "Property of the user of modifying the nickname" 
      description: "Modify it's own nickname"
      requestBody:
        required: true 
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Complete_user" }

      responses:
        "200":
          description: Nickname updated succesfully, returned the content 
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Complete_user' }

        "400":
          description: invalid object 

        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '500': 
          $ref: "#/components/responses/Server_error" 
  /users/{userid}:    
    parameters:
      - $ref: '#/components/parameters/user_id'
    
    get:
      operationId: getUserProfile
      tags: [user]   
      summary: get user profile 
      description:  get user profile by the unique identifier

      responses:
        "200":
          description: profile retrieved correctly 
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Complete_profile"}

        "400":  
          description: invalide identifier or not found a match 
        
        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '500': 
          $ref: "#/components/responses/Server_error"
...
  /users/{userid}/followers/:
    parameters:
     - $ref: '#/components/parameters/user_id'

    get:
      operationId: getFollowersList
      tags: [followers]
      summary: get user followers
      description: Get the list of followers for a user.
      responses:
        "200":
          # OK
          description: List of user followers retrieved successfully.
          content:
            application/json:
              schema:
                description: Users list.
                type: array
                minItems: 0
                maxItems: 99
                items: { $ref: '#/components/schemas/Complete_user' }
        "400":
          description: bad request 
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/Server_error'
  /users/{userid}/followings/:
    parameters:
      - $ref: '#/components/parameters/user_id'

    get:
      operationId: getFollowingsList
      tags: [following]
      summary: get user followings
      description: Get the list of users that a user is following.
      responses:
        "200":
          description: List of user followings retrieved successfully.
          content:
            application/json:
              schema:
                description: Users list.
                type: array
                minItems: 0
                maxItems: 99
                items: { $ref: '#/components/schemas/Complete_user' }
        "400":
          description: bad request 

        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/Server_error'

###########    
#############
...
  /users/{userid}/photos/{photoid}:  
    parameters:
      - $ref: '#/components/parameters/user_id'
      - $ref: '#/components/parameters/photo_id'
      

    get:
      operationId: getPhoto
      tags: ["photo"]
      summary: Downloads a photo
      description: user accesses its specific photo or other peculiar photo
      

      responses:
        '200':
          description: specific photo for a specific user retrieved
          content:
            image/*:
              schema: { $ref: "#/components/schemas/Raw_photo" }

        '400':
          description: invalid parameters 

        '401':
          $ref: "#/components/responses/Unauthorized"
          
          
        '500': 
          $ref: "#/components/responses/Server_error"

    delete:
      operationId: deletePhoto
      tags: [photo]
      summary: delete a pic
      description: user deletes its photo from the stream along with likes and comments 
      responses:
        "204":
          description: User deleted photo.

        "400":
          description: Invalid identifier parameter

        "401":
          $ref: '#/components/responses/Unauthorized'

        "403":
          description:  user not author of photo.

        "500":
          $ref: '#/components/responses/Server_error'
############

###### 
...
  /photos/{photoid}/comments/{commentid}:
    parameters:
      - $ref: '#/components/parameters/photo_id'
      - $ref: '#/components/parameters/comment_id'
    
    delete:
      operationId: uncommentPhoto
      tags: ['comment']
      summary: "deleting a comment"
      description: removing a comment from a specific post
      
      
      responses:
        "204":
          description: Comment successfully removed from the specified photo.
        "400":
          description: bad request 

        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '403': 
          description: requesting user not author of comment 
          
        '500': 
          $ref: "#/components/responses/Server_error"
#######

 ###########     
  /like/{photoid}:
    parameters:
      - $ref: '#/components/parameters/photo_id'

    
    put: 
      operationId: likePhoto
      tags: ["like"]
      summary: "put a like"
      description: through the identifier of a pic I put a like 
      
      responses:

        "204":
          description: The like has been put to the specified photo.

        "400":
          description: invalid parameter

        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '500': 
          $ref: "#/components/responses/Server_error"
    
    delete:
      operationId: unlikePhoto
      tags: ["like"]
      summary: "unlike a photo" 
      description: "user delete a like from a photo"
      
  
      responses: 
        "204":
          description: The like has been removed from the specified photo.

        "400":
          description: invalid parameter

        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '404': 
          $ref: "#/components/responses/Not_found"
          
        '500': 
          $ref: "#/components/responses/Server_error"
 #########     
components :
  securitySchemes:
      bearerAuth:            
        type: http
        scheme: bearer
        bearerFormat: JWT

    
      
#################################
  parameters: 
      user_id:
        name: userid 
        in: path 
        required: true
        description: identifier of the user 
        schema:
          $ref: "#/components/schemas/User_login"
      photo_id:
        name: photoid
        in: path 
        required: true
        description: identifier of the photo
        schema:
          $ref: "#/components/schemas/Photo_id"
...
      comment_id:
        name: commentid
        in: path
        required: true
        description: comment identifier 
        schema:
          $ref: "#/components/schemas/Comment_identifier"

           
#################################    
  schemas:
      User_login: # USER_ID 
        description: Identifier sent by user during the login
        readOnly: true 
        type: number  
        format: uint64
        example: 24
      Nickname: # USERNAME 
        description: Username 
        type: string 
        pattern: '^.*?$'
        minLength: 3
        maxLength: 16
        example: "Annafabulous"
      Complete_user: #USER 
        description: Everything you know about a user ID and Nickname
        type: object 
        properties:
              user_login:
                $ref: "#/components/schemas/User_login"   
              nickname: 
                $ref: "#/components/schemas/Nickname"   
      Photo_id: #PHOTOID
        description: unique identifier of a picture
        readOnly: True 
        type: number 
        format: uint64
        example: 145
      Comment_identifier: #COMMENTID
        description: identifier of the comment 
        readOnly: true 
        type: number
        format: uint64
      Date:
        description: realease date of a comment or a pic 
        type: string
        format: date-time
        minLength: 19
        maxLength: 19
      Raw_photo: #BINARYIMAGE
        description: Image encoding 
        type: string
        minLength: 4
        maxLength: 65535
        format: binary 
      Comment_text:
            description: text of the comment 
            type: string
            pattern: '^.*?$'
            minLength: 1
            maxLength: 2200
            example: nice pic 
      Comment: 
        description: comment of a specific User on a specific Photo 
        type: object 
        properties: 
          user_id:
            $ref: "#/components/schemas/User_login"             
          nickname: 
            $ref: "#/components/schemas/Nickname" 
          comment_id:
            $ref: "#/components/schemas/Comment_identifier"  
          photo_id:
            $ref: "#/components/schemas/Photo_id"
          text:
            $ref: "#/components/schemas/Comment_text" 
          date:
            $ref: "#/components/schemas/Date"
      Photo:
        description: Contains Everything of a photo
        type: object
        properties:
          photo_id:
            $ref: "#/components/schemas/Photo_id"
          owner:
            $ref: "#/components/schemas/User_login"
          format:
            description: the format of the photo 
            type: string 
            enum:
              - png
              - jpg 
            pattern: '^.*?$'
            minLength: 3
            maxLength: 3  
          comments:
            description: array of comments in a single photo
            type: array 
            minItems: 0
            maxItems: 99
            items: {$ref: "#/components/schemas/Comment" }                      
          likes: 
            description: array of users that liked the picture 
            type: array 
            minItems: 0
            maxItems: 99
            items: { $ref: "#/components/schemas/Complete_user" }                     
          date:
            $ref: "#/components/schemas/Date"
      Complete_profile: #PROFILE 
        description: Object containing all the profile info
        type: object
        properties:

          nickname: 
            $ref: "#/components/schemas/Nickname"
          
          #mi piacerebbe metter che invece del numero di foto ---> mettere numero likes 
          likesCount:
            description: total number of likes for all its pictures 
            type: number 
            format: uint64 

          #ce lo metto per sicurezza  
          photosCount:
            description: Total number of photos.
            type: number
            format: uint64

          followers:
            description: Array of followers 
            type: array 
            minItems: 0
            maxItems: 100
            items: { $ref: "#/components/schemas/Nickname" }  #### VEDERE POI SE FUNZIONA ( PRIMA ERA COMPLETE_USER)

          
          following:
            description: Array of followed users
            type: array 
            minItems: 0
            maxItems: 100
            items: { $ref: "#/components/schemas/Nickname" } 
              
          inBannedList:
            description: user is in the banned list 
            type: boolean

          meBanned:
            description: I am in the Banned list of a user 
            type: boolean 

                    ##### aggiunta mia: nel profilo compelto voglio far vedere anche le foto 
          photos:
            description: Array of photos 
            type: array
            minItems: 0
            maxItems: 50 
            items: { $ref: "#/components/schemas/Photo" }

#################################    
  responses : 
      Unauthorized:
        description: "The user is not logged in"   
      Not_found:
        description: User not found
      Server_error:
        description: Server error
      Forbidden:
        description: user in banned list
...
security:
  - bearerAuth: []